+++ src/api/api.ts
@@ -11,10 +11,12 @@ import Toot, { earliestTootedAt } from './objects/toot';
 import { buildAccountNames } from "./objects/account";
 import { countValues, sortKeysByValue } from '../helpers';
 import { Key, StorageKey, StorageValue, StringNumberDict, TimelineData, UserData, WeightName} from "../types";
+import { repairTag } from "./objects/tag";
 
 type ApiMutex = Record<StorageKey, Mutex>;
 
 export const INSTANCE = "instance"
+export const LINKS = "links";
 export const STATUSES = "statuses"
 export const TAGS = "tags"
 
@@ -78,8 +80,8 @@ export class MastoApi {
         // Only retrieve trending toots on the first call to this method
         if (!maxId) {
             promises = promises.concat([
-                MastodonServer.fediverseTrendingToots(),
                 fetchRecentTootsForTrendingTags(),
+                MastodonServer.fediverseTrendingToots(),
             ]);
         }
 
@@ -105,7 +107,7 @@ export class MastoApi {
 
         return {
             followedAccounts: buildAccountNames(responses[0]),
-            followedTags: countValues<mastodon.v1.Tag>(responses[3], (tag) => tag.name.toLowerCase()),
+            followedTags: countValues<mastodon.v1.Tag>(responses[3], (tag) => tag.name),
             mutedAccounts: buildAccountNames(responses[1].concat(responses[2])),
             serverSideFilters: responses[4],
         } as UserData;
@@ -179,10 +181,12 @@ export class MastoApi {
 
     // Get hashtags the user is following
     async getFollowedTags(): Promise<mastodon.v1.Tag[]> {
-        return await this.fetchData<mastodon.v1.Tag>({
+        const followedTags = await this.fetchData<mastodon.v1.Tag>({
             fetch: this.api.v1.followedTags.list,
             label: WeightName.FOLLOWED_TAGS
         });
+
+        return followedTags.map(repairTag);
     }
 
     // Get the user's recent notifications