+++ src/scorer/feature/followed_tags_scorer.ts
@@ -11,16 +11,18 @@ import { MastoApi } from '../../api/api';
 import { StringNumberDict, WeightName } from '../../types';
 
 
-export default class FollowedTagsFeatureScorer extends FeatureScorer {
+export default class FollowedTagsScorer extends FeatureScorer {
     constructor() {
         super(WeightName.FOLLOWED_TAGS);
     }
 
+    // Return a dict keyed by tag name (values should be all 1)
     async featureGetter(): Promise<StringNumberDict> {
-        const tags = await MastoApi.instance.getFollowedTags();
-        return countValues<mastodon.v1.Tag>(tags, (tag) => tag.name?.toLowerCase());
+        const followedTags = await MastoApi.instance.getFollowedTags();
+        return countValues<mastodon.v1.Tag>(followedTags, tag => tag.name);
     }
 
+    // Sets the followedTags property on the Toot object before returning the score
     async _score(toot: Toot) {
         toot.followedTags = toot.tags.filter((tag) => tag.name in this.requiredData);
         return toot.followedTags.length;