+++ src/types.ts
@@ -40,6 +40,7 @@ export enum WeightName {
     NUM_RETOOTS = 'NumRetoots',
     RETOOTED_IN_FEED = 'RetootedInFeed',
     TIME_DECAY = 'TimeDecay',
+    TRENDING_LINKS = 'TrendingLinks',
     TRENDING_TAGS = "TrendingTags",
     TRENDING_TOOTS = "TrendingToots",
     VIDEO_ATTACHMENTS = 'VideoAttachments',
@@ -79,6 +80,7 @@ export type Config = {
     // Trending tags
     numTootsPerTrendingTag: number;
     numDaysToCountTrendingTagData: number;
+    numTrendingLinksPerServer: number;
     numTrendingTags: number;
     numTrendingTagsPerServer: number;
     numTrendingTagsToots: number;
@@ -128,8 +130,10 @@ export type UserData = {
     followedTags: StringNumberDict,
     mutedAccounts: AccountNames,
     serverSideFilters: mastodon.v2.Filter[],
+    // trendingLinks: mastodon.v1.TrendLink[],
 };
 
+// Serialized version of a Toot
 export interface TootExtension extends mastodon.v1.Status {
     followedTags?: mastodon.v1.Tag[];  // Array of tags that the user follows that exist in this toot
     isFollowed?: boolean;              // Whether the user follows the account that posted this toot
@@ -137,6 +141,7 @@ export interface TootExtension extends mastodon.v1.Status {
     reblogBy?: mastodon.v1.Account;    // The account that retooted this toot (if any)
     scoreInfo?: TootScore;             // Scoring info for weighting/sorting this toot
     trendingRank?: number;             // Most trending on a server gets a 10, next is a 9, etc.
+    trendingLinks?: TrendingLink[];    // Links that are trending in this toot
     trendingTags?: TrendingTag[];      // Tags that are trending in this toot
 };
 
@@ -148,15 +153,22 @@ export type TootScore = {
     weightedScores: Weights;
 };
 
+export interface TrendingLink extends mastodon.v1.TrendLink {
+    numToots?: number;
+    numAccounts?: number;
+};
+
 export interface TrendingTag extends mastodon.v1.Tag {
     numAccounts?: number;
     numToots?: number;
     trendingRank?: number;
 };
 
+export type TrendingWithHistory = TrendingLink | TrendingTag;
+
 export type StorageValue = FeedFilterSettings | FeedFilterSettingsSerialized | StringNumberDict |
     TootExtension[] | TootURIs | Weights | mastodon.v1.Account | mastodon.v1.Account[] | mastodon.v2.Filter[] |
-    number;
+    mastodon.v1.TrendLink[] | number;
 
 
 // From https://dev.to/nikosanif/create-promises-with-timeout-error-in-typescript-fmm