+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## NEXT
+* Add `TrendingLinksScorer` to boost toots that contain Fediverse-wide trending links
+* `TrendingTagsScorer` uses max of `numAccounts` not `log2(sum(numAccounts))`
 * Add blocked accounts to muted account list
 * [bug] Add `@server.com` string to `Account` objects for users on the home server that didn't have it.